Work with newer GnuPG::Interface versions again
Newer GnuPG::Interface versions are more picky when it comes to command
and command_args.  Previously, they would accept things in any order and
would not care which of the two arrays had which.  Now, command_args is
really limited to have arguments for the command listed in command.  Any
options go with, and in front of, the command.